Add InfinixOS nougat source codes
infinixOS, named "XOS Chameleon", an Updated version of XUI from Infinix Mobiles.

This is the official source code for "Infinix Hot 4 Pro" based on Android Nougat.

The halogenOS XOS is Not the same. And they are built for many Qualcomm-based chipset.
But the XOS Chameleon is only built for MediaTek-based Infinix Mobiles.